Update to new moose game theme and centralize theme and sprite loading

+++ lib/screens/moose_game/sprite.dart
@@ -1,7 +1,8 @@
 class Sprite {
   final String imagePath;
-  final int imageWidth;
-  final int imageHeight;
+  // Allow fractional sizes so we can scale graphics to exactly line up on a pixel level
+  final double imageWidth;
+  final double imageHeight;
 
   const Sprite(this.imagePath, this.imageWidth, this.imageHeight);
 }
